+++ R/meta.R
@@ -7,7 +7,8 @@
 #' @importFrom googleAuthR gar_api_generator
 #' 
 #' @export
-google_analytics_meta <- function(){
+#' 
+ga_meta <- function(){
   
   meta <- gar_api_generator("https://www.googleapis.com/analytics/v3",
                             "GET",