+++ R/options.R
@@ -9,11 +9,11 @@
     googleAuthR.webapp.client_secret = "0zBtmZ_klTEzXUaTUTP5AkNQ",
     googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/analytics", 
                                     "https://www.googleapis.com/auth/analytics.readonly",
-                                    "https://www.googleapis.com/auth/analytics.manage.users.readonly",
                                     "https://www.googleapis.com/auth/analytics.edit",
                                     "https://www.googleapis.com/auth/analytics.manage.users",
                                     "https://www.googleapis.com/auth/analytics.user.deletion"),
-    googleAuthR.httr_oauth_cache = "ga.oauth"
+    googleAuthR.httr_oauth_cache = "ga.oauth",
+    googleAuthR.quotaUser = Sys.info()[["user"]]
   )
   
   toset <- !(names(op.googleAnalyticsR) %in% names(op))
@@ -24,24 +24,6 @@
   options(
     googleAuthR.batch_endpoint = "https://www.googleapis.com/batch/analytics/v3",
     googleAuthR.tryAttempts = 1)
-  
-  if(Sys.getenv("GA_CLIENT_ID") != ""){
-    options(googleAuthR.client_id = Sys.getenv("GA_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_CLIENT_SECRET") != ""){
-    options(googleAuthR.client_secret = Sys.getenv("GA_CLIENT_SECRET"))
-  }
-  
-  if(Sys.getenv("GA_WEB_CLIENT_ID") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_WEB_CLIENT_SECRET") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_SECRET"))
-  }
-  
-  default_project_message()
   
   f <- function(req){
     
@@ -71,24 +53,19 @@
 
 .onAttach <- function(libname, pkgname){
   
-  if(Sys.getenv("GA_CLIENT_ID") != ""){
-    options(googleAuthR.client_id = Sys.getenv("GA_CLIENT_ID"))
+  if(Sys.getenv("GAR_CLIENT_JSON") != ""){
+    googleAuthR::gar_set_client(json = Sys.getenv("GAR_CLIENT_JSON"))
   }
   
-  if(Sys.getenv("GA_CLIENT_SECRET") != ""){
-    options(googleAuthR.client_secret = Sys.getenv("GA_CLIENT_SECRET"))
-  }
+  needed <- c("https://www.googleapis.com/auth/analytics", 
+              "https://www.googleapis.com/auth/analytics.readonly",
+              "https://www.googleapis.com/auth/analytics.edit",
+              "https://www.googleapis.com/auth/analytics.manage.users",
+              "https://www.googleapis.com/auth/analytics.user.deletion")
   
-  if(Sys.getenv("GA_WEB_CLIENT_ID") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_WEB_CLIENT_SECRET") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_SECRET"))
-  }
-  
-  needed <- c("https://www.googleapis.com/auth/analytics.readonly")
-  
+  googleAuthR::gar_attach_auto_auth(needed, 
+                                    environment_var = "GARGLE_EMAIL")
+  # for json files
   googleAuthR::gar_attach_auto_auth(needed, 
                                     environment_var = "GA_AUTH_FILE")